requirements.txt
numpy>=1.23
opencv-python>=4.8
torch>=2.0
torchvision>=0.15
tqdm>=4.65
PyYAML>=6.0
transformers>=4.32
huggingface-hub>=0.17
